Changes external particle index to size_t type

+++ applications/ionMobility/BT-RS-DMSSim/BT-RS-DMSSim.cpp
@@ -236,7 +236,7 @@ int main(int argc, const char * argv[]) {
                 particle->setLocation(initialPositions[k]);
 
                 particlesPtrs.push_back(particle.get());
-                rsSim.addParticle(particle.get(), static_cast<int>(nParticlesTotal));
+                rsSim.addParticle(particle.get(), nParticlesTotal);
                 particles.push_back(std::move(particle));
                 trajectoryAdditionalParams.push_back(std::vector<double>(1));
                 nParticlesTotal++;
@@ -391,7 +391,7 @@ int main(int argc, const char * argv[]) {
                 reactionConditions.electricField = fieldMagnitude;
                 reactionConditions.temperature = backgroundTemperatureFct(particles[i]->getLocation());
 
-                bool reacted = rsSim.react(static_cast<int>(i), reactionConditions, dt_s);
+                bool reacted = rsSim.react(i, reactionConditions, dt_s);
                 int substIndex = substanceIndices.at(particles[i]->getSpecies());
                 particles[i]->setFloatAttribute(key_ChemicalIndex, substIndex);
